Code Restructuration
The part of the methods and linear equations systems solutions were refactored to a directory called domain, because is the core of the project
