+++ BACKEND/app/tests/domain/methods/test.py
@@ -1,4 +1,4 @@
-from app.utils.methods import bisection, false_rule, fixed_point, newton_raphson, secant, first_modified_newton_method, second_modified_newton_method
+from app.domain.methods import bisection, false_rule, fixed_point, newton_raphson, secant, first_modified_newton_method, second_modified_newton_method
 from app.utils.utils import parse_expression
 from app.routes.routes import logger
 from fastapi.exceptions import HTTPException
